+++ src/components/Alerts/OcclusionMaskModal/OcclusionMaskModal.tsx
@@ -7,20 +7,22 @@ import {
   DialogTitle,
   Typography,
 } from '@mui/material';
-import { useCallback, useEffect, useState } from 'react';
+import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 
-import { type DetectionType, getDetectionsBySequence } from '@/services/alerts';
+import { getDetectionsBySequence } from '@/services/alerts';
 import {
   addOcclusionMask,
-  type BboxType,
   clearOcclusionMasks,
+  getOcclusionMasks,
+} from '@/services/occlusionMasks';
+import type { SequenceWithCameraInfoType } from '@/utils/alerts';
+import {
+  type BboxType,
   enlargeBbox,
   getHighestConfidenceBbox,
   getHighestConfidenceDetection,
   getNonOverlappingMasks,
-  getOcclusionMasks,
-} from '@/services/occlusionMasks';
-import type { SequenceWithCameraInfoType } from '@/utils/alerts';
+} from '@/utils/occlusionMasks';
 import { useTranslationPrefix } from '@/utils/useTranslationPrefix';
 
 interface BboxOverlayProps {
@@ -62,8 +64,90 @@ export const OcclusionMaskModal = ({
   sequence,
 }: OcclusionMaskModalProps) => {
   const { t } = useTranslationPrefix('alerts.occlusionMask');
-  const [detectionsData, setDetectionsData] = useState<DetectionType[]>([]);
-  const [existingMasks, setExistingMasks] = useState<BboxType[]>([]);
+  const queryClient = useQueryClient();
+
+  // Query for detections data
+  const {
+    data: detectionsData = [],
+    isLoading: isLoadingDetections,
+    error: detectionsError,
+  } = useQuery({
+    queryKey: ['detections', sequence.id],
+    queryFn: () => getDetectionsBySequence(sequence.id),
+    enabled: open && !!sequence.id,
+  });
+
+  // Query for existing occlusion masks
+  const { data: existingMasks = [], isLoading: isLoadingMasks } = useQuery({
+    queryKey: [
+      'occlusionMasks',
+      sequence.camera?.name,
+      sequence.camera?.angle_of_view,
+    ],
+    queryFn: () => {
+      if (!sequence.camera?.name || sequence.camera.angle_of_view === null) {
+        return [];
+      }
+      const masks = getOcclusionMasks(
+        sequence.camera.name,
+        sequence.camera.angle_of_view
+      );
+      return getNonOverlappingMasks(masks);
+    },
+    enabled:
+      open && !!sequence.camera?.name && sequence.camera.angle_of_view !== null,
+  });
+
+  // Mutation for adding occlusion mask
+  const addMaskMutation = useMutation({
+    mutationFn: ({
+      cameraName,
+      angleOfView,
+      bbox,
+    }: {
+      cameraName: string;
+      angleOfView: number;
+      bbox: BboxType;
+    }) => {
+      addOcclusionMask(cameraName, angleOfView, bbox);
+      return Promise.resolve();
+    },
+    onSuccess: () => {
+      // Invalidate and refetch occlusion masks
+      void queryClient.invalidateQueries({
+        queryKey: [
+          'occlusionMasks',
+          sequence.camera?.name,
+          sequence.camera?.angle_of_view,
+        ],
+      });
+      onClose();
+    },
+  });
+
+  // Mutation for clearing all masks
+  const clearMasksMutation = useMutation({
+    mutationFn: ({
+      cameraName,
+      angleOfView,
+    }: {
+      cameraName: string;
+      angleOfView: number;
+    }) => {
+      clearOcclusionMasks(cameraName, angleOfView);
+      return Promise.resolve();
+    },
+    onSuccess: () => {
+      // Invalidate and refetch occlusion masks
+      void queryClient.invalidateQueries({
+        queryKey: [
+          'occlusionMasks',
+          sequence.camera?.name,
+          sequence.camera?.angle_of_view,
+        ],
+      });
+    },
+  });
 
   // Derive highestDetection and proposedBbox from detectionsData
   const highestDetection = getHighestConfidenceDetection(detectionsData);
@@ -74,49 +158,39 @@ export const OcclusionMaskModal = ({
       })()
     : null;
 
-  const loadDetections = useCallback(async () => {
-    if (!sequence.camera?.name || sequence.azimuth === null) {
-      return;
-    }
-
-    try {
-      const detections = await getDetectionsBySequence(sequence.id);
-      setDetectionsData(detections);
-
-      // Load existing occlusion masks from localStorage
-      const masks = getOcclusionMasks(sequence.camera.name, sequence.azimuth);
-      const nonOverlappingMasks = getNonOverlappingMasks(masks);
-      setExistingMasks(nonOverlappingMasks);
-    } catch (err) {
-      console.error('Error loading detections:', err);
-    }
-  }, [sequence]);
-
-  useEffect(() => {
-    if (open) {
-      void loadDetections();
-    }
-  }, [open, loadDetections]);
-
   const handleConfirmSelection = () => {
-    if (!proposedBbox || !sequence.camera?.name || sequence.azimuth === null) {
+    if (
+      !proposedBbox ||
+      !sequence.camera?.name ||
+      sequence.camera.angle_of_view === null
+    ) {
       return;
     }
 
-    addOcclusionMask(sequence.camera.name, sequence.azimuth, proposedBbox);
-    onClose();
+    addMaskMutation.mutate({
+      cameraName: sequence.camera.name,
+      angleOfView: sequence.camera.angle_of_view,
+      bbox: proposedBbox,
+    });
   };
 
   const handleDeleteAll = () => {
-    if (!sequence.camera?.name || sequence.azimuth === null) {
+    if (!sequence.camera?.name || sequence.camera.angle_of_view === null) {
       return;
     }
 
-    clearOcclusionMasks(sequence.camera.name, sequence.azimuth);
-    setExistingMasks([]);
+    clearMasksMutation.mutate({
+      cameraName: sequence.camera.name,
+      angleOfView: sequence.camera.angle_of_view,
+    });
   };
 
   const imageUrl = highestDetection?.url ?? '';
+  const isLoading = isLoadingDetections || isLoadingMasks;
+
+  if (detectionsError) {
+    console.error('Error loading detections:', detectionsError);
+  }
 
   return (
     <Dialog
@@ -140,59 +214,75 @@ export const OcclusionMaskModal = ({
           <Typography variant="body1" gutterBottom>
             {t('explanationText')}
           </Typography>
-          {/* <Typography variant="body2" color="text.secondary" gutterBottom>
-            {t('explanationText')}
-          </Typography> */}
 
-          {imageUrl && (
+          {isLoading ? (
             <Box
-              position="relative"
               sx={{
-                mt: 2,
-                border: 1,
-                borderColor: 'divider',
-                borderRadius: 1,
-                overflow: 'hidden',
-                display: 'inline-block',
-                maxWidth: '100%',
+                display: 'flex',
+                justifyContent: 'center',
+                alignItems: 'center',
+                minHeight: '200px',
               }}
             >
-              <img
-                src={imageUrl}
-                alt="Detection"
-                style={{
-                  width: '100%',
-                  height: 'auto',
-                  display: 'block',
-                }}
-              />
-
-              {/* Existing occlusion masks (red) */}
-              {existingMasks.map((mask) => (
-                <BboxOverlay
-                  key={`mask-${mask.xmin}-${mask.ymin}-${mask.xmax}-${mask.ymax}`}
-                  bbox={mask}
-                  color="red"
-                />
-              ))}
-
-              {/* Proposed new mask (green) */}
-              {proposedBbox && (
-                <BboxOverlay bbox={proposedBbox} color="green" />
-              )}
+              <Typography variant="body2" color="text.secondary">
+                Loading...
+              </Typography>
             </Box>
+          ) : (
+            imageUrl && (
+              <Box
+                position="relative"
+                sx={{
+                  mt: 2,
+                  border: 1,
+                  borderColor: 'divider',
+                  borderRadius: 1,
+                  overflow: 'hidden',
+                  display: 'inline-block',
+                  maxWidth: '100%',
+                }}
+              >
+                <img
+                  src={imageUrl}
+                  alt="Detection"
+                  style={{
+                    width: '100%',
+                    height: 'auto',
+                    display: 'block',
+                  }}
+                />
+
+                {/* Existing occlusion masks (red) */}
+                {existingMasks.map((mask) => (
+                  <BboxOverlay
+                    key={`mask-${mask.xmin}-${mask.ymin}-${mask.xmax}-${mask.ymax}`}
+                    bbox={mask}
+                    color="red"
+                  />
+                ))}
+
+                {/* Proposed new mask (green) */}
+                {proposedBbox && (
+                  <BboxOverlay bbox={proposedBbox} color="green" />
+                )}
+              </Box>
+            )
           )}
         </Box>
       </DialogContent>
       <DialogActions sx={{ p: 3, gap: 1 }}>
-        <Button
-          onClick={handleDeleteAll}
-          color="error"
-          variant="outlined"
-          disabled={existingMasks.length === 0}
-        >
-          {t('deleteAllButton')}
-        </Button>
+        {existingMasks.length > 0 && (
+          <Button
+            onClick={handleDeleteAll}
+            color="error"
+            variant="outlined"
+            disabled={clearMasksMutation.isPending}
+          >
+            {clearMasksMutation.isPending
+              ? 'Deleting...'
+              : t('deleteAllButton')}
+          </Button>
+        )}
         <Box sx={{ flex: 1 }} />
         <Button onClick={onClose} color="inherit">
           {t('cancelButton')}
@@ -201,9 +291,9 @@ export const OcclusionMaskModal = ({
           onClick={handleConfirmSelection}
           color="success"
           variant="contained"
-          disabled={!proposedBbox}
+          disabled={!proposedBbox || addMaskMutation.isPending || isLoading}
         >
-          {t('confirmButton')}
+          {addMaskMutation.isPending ? 'Adding...' : t('confirmButton')}
         </Button>
       </DialogActions>
     </Dialog>
