+++ src/locales/es/translation.json
@@ -95,7 +95,7 @@
       "title": "Gestionar máscaras de oclusión",
       "loading": "Cargando datos de detección...",
       "confirmZoneText": "Confirmar la zona de oclusión",
-      "explanationText": "Las máscaras de oclusión ocultan áreas que generan falsos positivos recurrentes. La zona verde muestra la nueva máscara propuesta basada en la detección con mayor confianza (ampliada en un 10%). Las zonas rojas muestran las máscaras existentes.",
+      "explanationText": "Las máscaras de oclusión ocultan áreas que generan falsos positivos recurrentes. La zona verde muestra la nueva máscara propuesta. Las zonas rojas muestran las máscaras existentes.",
       "confirmButton": "Confirmar selección",
       "deleteAllButton": "Eliminar zonas de oclusión",
       "cancelButton": "Cancelar",
